requirements.txt
alembic==1.2.1
appnope==0.1.0
backcall==0.1.0
beautifulsoup4==4.8.0
Bootstrap-Flask==1.1.0
certifi==2019.6.16
chardet==3.0.4
Click==7.0
convertdate==2.2.0
cycler==0.10.0
Cython==0.29.14
decorator==4.4.0
dnspython==1.16.0
dominate==2.4.0
email-validator==1.0.4
fbprophet==0.5
findspark==1.3.0
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
geographiclib==1.50
geopy==1.20.0
holidays==0.9.12
idna==2.8
imageio==2.6.1
ipykernel==5.1.2
ipyparallel==6.2.4
ipython==7.8.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.3
jupyter-client==5.3.1
jupyter-core==4.5.0
kiwisolver==1.1.0
lunardate==0.2.0
Mako==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.2
metakernel==0.24.2
mysql-connector-python==8.0.11
networkx==2.4
numpy==1.17.4
opencv-contrib-python==4.1.2.30
opencv-python==4.1.2.30
pandas==0.25.3
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.1
pipenv==2018.11.26
plotly==4.4.1
prompt-toolkit==2.0.9
protobuf==3.9.1
psycopg2==2.8.4
ptyprocess==0.6.0
Pygments==2.4.2
PyMeeus==0.3.6
pyparsing==2.4.5
PyQt5==5.13.2
PyQt5-sip==4.19.19
pystan==2.19.1.1
python-dateutil==2.8.0
python-editor==1.0.4
python-zillow==0.2.0
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.1.2
pyzillow==0.5.5
pyzmq==18.1.0
requests==2.22.0
retrying==1.3.3
scikit-image==0.16.2
scipy==1.4.0
setuptools-git==1.2
six==1.12.0
soupsieve==1.9.3
spylon==0.3.0
spylon-kernel==0.4.1
SQLAlchemy==1.3.10
tornado==6.0.3
traitlets==4.3.2
urllib3==1.25.3
virtualenv==16.7.7
virtualenv-clone==0.5.3
visitor==0.1.3
wcwidth==0.1.7
Werkzeug==0.16.0
WTForms==2.2.1
xmltodict==0.12.0
